prescriptions2/requirements.txt
Django~=3.2.6
Pika==1.2.0
djangorestframework==3.12.4
psycopg2==2.9.1
django-cors-headers==3.8.0
psycopg2-binary>=2.9.1
python-decouple==3.4
